requirements.txt
asgiref==3.3.4
boto3==1.17.48
botocore==1.20.48
Django==3.2
django-cors-headers==3.7.0
django-ninja==0.12.2
django-storages==1.11.1
jmespath==0.10.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pydantic==1.7.3
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
